Move interfaces into digdag-spi
