+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/DummyParamServerClient.java
@@ -2,6 +2,8 @@ package io.digdag.standards.operator.param;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.base.Optional;
+import io.digdag.spi.ParamServerClient;
+import io.digdag.spi.ParamServerClientConnection;
 import io.digdag.spi.Record;
 
 import java.util.function.Consumer;
