+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/DummyServerClientConnection.java
@@ -1,5 +1,7 @@
 package io.digdag.standards.operator.param;
 
+import io.digdag.spi.ParamServerClientConnection;
+
 public class DummyServerClientConnection
         implements ParamServerClientConnection
 {
