+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/DummyServerClientConnectionManager.java
@@ -1,5 +1,8 @@
 package io.digdag.standards.operator.param;
 
+import io.digdag.spi.ParamServerClientConnection;
+import io.digdag.spi.ParamServerClientConnectionManager;
+
 public class DummyServerClientConnectionManager
         implements ParamServerClientConnectionManager
 {
