+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/ParamServerModule.java
@@ -3,6 +3,7 @@ package io.digdag.standards.operator.param;
 import com.google.inject.Binder;
 import com.google.inject.Module;
 import com.google.inject.Scopes;
+import io.digdag.spi.ParamServerClientConnectionManager;
 
 public class ParamServerModule
         implements Module
