+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/PostgresqlServerClientConnection.java
@@ -1,5 +1,6 @@
 package io.digdag.standards.operator.param;
 
+import io.digdag.spi.ParamServerClientConnection;
 import org.skife.jdbi.v2.Handle;
 
 public class PostgresqlServerClientConnection implements ParamServerClientConnection<Handle>
