+++ digdag-standards/src/main/java/io/digdag/standards/operator/param/RedisServerClientConnection.java
@@ -1,5 +1,6 @@
 package io.digdag.standards.operator.param;
 
+import io.digdag.spi.ParamServerClientConnection;
 import redis.clients.jedis.Jedis;
 
 public class RedisServerClientConnection implements ParamServerClientConnection<Jedis>
